% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monolith.r
\name{asset_query_maven}
\alias{asset_query_maven}
\title{Query Maven Assets}
\usage{
asset_query_maven(
  artifactId,
  repository = "maven-releases",
  group = "",
  version = "",
  maven.extension = "jar",
  maven.classifier = "",
  url = "https://search.maven.org/solrsearch/select"
)
}
\arguments{
\item{artifactId}{The Maven artifact ID.}

\item{repository}{The Maven repository.}

\item{group}{The group the asset belongs to.}

\item{maven.extension}{The asset file extension.}

\item{maven.classifier}{The asset classifier.}

\item{url}{The repository base URL.}
}
\value{
A data frame of asset information.
}
\description{
Query assets from a Maven repository.
}
\keyword{internal}
