% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monolith.r
\name{asset_query_nexus}
\alias{asset_query_nexus}
\title{Query Nexus Assets}
\usage{
asset_query_nexus(
  artifactId,
  repository = "maven-releases",
  group = "mil.army.usace.hec",
  version = "",
  maven.extension = "jar",
  maven.classifier = "",
  url = "https://www.hec.usace.army.mil/nexus/service/rest/v1/search/assets"
)
}
\arguments{
\item{artifactId}{The Maven artifact ID.}

\item{repository}{The Maven repository.}

\item{group}{The group the asset belongs to.}

\item{maven.extension}{The asset file extension.}

\item{maven.classifier}{The asset classifier.}

\item{url}{The repository base URL.}
}
\value{
A data frame of asset information.
}
\description{
Query assets from a Nexus repository.
}
\keyword{internal}
