% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.r
\name{build_fields_table}
\alias{build_fields_table}
\alias{build_methods_table}
\title{Fields and Methods Data Frame}
\usage{
build_fields_table(jObject)

build_methods_table(jObject)
}
\arguments{
\item{jObject}{A Java object.}
}
\value{
A dataframe with fields "FULLNAME", "SHORTNAME", "CLASS",
"SIGNATURE", and (for methods) "ARGUMENTS".
}
\description{
Get a data frame of Java object fields or methods.
}
\keyword{internal}
