% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.r
\name{.catalog_store}
\alias{.catalog_store}
\title{DSS Catalog store}
\usage{
.catalog_store(condensed)
}
\arguments{
\item{condensed}{If \code{TRUE}, create a store that calls the condensed
catalog methods.}
}
\description{
The \code{dssrip2} catalog store. Caches the catalogs of DSS files.
}
\details{
The file store stores a list of \code{hec.heclib.dss.HecDss}
Java object references named according to the normalized file
path of the originating DSS file. The file store provides four
methods:
\itemize{
\item \code{get()}: Retrieve a DSS catalog from the store.
\item \code{set()}: Create a DSS catalog and add it to the store.
\item \code{drop()}: Remove a DSS catalog from the store.
\item \code{list()}: List the DSS catalog names in the store.
If the \code{set()} method fails with a message that \code{javaHeclib.dll}
cannot be found, this usually indicates that \code{dss.home} is not set
correctly.
}
}
\keyword{internal}
