% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.r
\name{.file_store}
\alias{.file_store}
\title{DSS File Store}
\usage{
.file_store()
}
\description{
The \code{dssrip2} file store. manages the Java DSS File objects.
}
\details{
The file store stores a list of \code{hec.heclib.dss.HecDss}
Java object references named according to the normalized file
path of the originating DSS file. The file store provides four
methods:
\itemize{
\item \code{get()}: Retrieve a DSS file handle from the store.
\item \code{set()}: Create a DSS file handle and add it to the store.
\item \code{drop()}: Remove a DSS file handle from the store.
\item \code{list()}: List the DSS file handle names in the store.
If the \code{set()} method fails with a message that \code{javaHeclib.dll}
cannot be found, this usually indicates that \code{dss.home} is not set
correctly.
}
}
\keyword{internal}
