% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java-utils.r
\name{.javaVectorToStrings}
\alias{.javaVectorToStrings}
\title{Java Vector To Strings}
\usage{
.javaVectorToStrings(vect)
}
\arguments{
\item{vect}{A java vector}
}
\value{
A character vector.
}
\description{
Convert java vectors returned by catalog functions to strings.
}
\keyword{internal}
