% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.r
\name{dss_attributes}
\alias{dss_attributes}
\alias{dss_add_attributes}
\title{DSS Attributes}
\usage{
dss_attributes(x)

dss_add_attributes(x, attributes_list = list())
}
\arguments{
\item{x}{An R object.}

\item{attributes_list}{A named list of DSS attributes.}
}
\value{
The object \code{x}, with additional attribute
\code{"dss_attributes"}.
}
\description{
Get and set DSS attributes.
}
