% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.r
\name{dss_attributes_grid}
\alias{dss_attributes_grid}
\title{DSS Grid Attributes}
\usage{
dss_attributes_grid(gridObj)
}
\arguments{
\item{gridObj}{A \code{hec.io.GridContainer} Java object reference.}
}
\value{
A data frame.
}
\description{
Get DSS grod attributes.
}
\keyword{internal}
