% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.r
\name{dss_attributes_paired}
\alias{dss_attributes_paired}
\title{DSS Paired Data Attributes}
\usage{
dss_attributes_paired(pdObj)
}
\arguments{
\item{pdObj}{A \code{hec.io.PairedDataContainer} Java object reference.}
}
\value{
A data frame.
}
\description{
Get DSS paired data attributes.
}
\keyword{internal}
