% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.r
\name{dss_attributes_timeseries}
\alias{dss_attributes_timeseries}
\title{DSS Time Series Attributes}
\usage{
dss_attributes_timeseries(tsObj)
}
\arguments{
\item{tsObj}{A \code{hec.io.TimeSeriesContainer} Java object reference.}
}
\value{
A data frame.
}
\description{
Get DSS time series attributes.
}
\keyword{internal}
