% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.r
\name{dss_catalog_condensed}
\alias{dss_catalog_condensed}
\alias{dss_catalog_full}
\title{DSS Catalog Retrieval}
\usage{
dss_catalog_condensed(filename, rebuild)

dss_catalog_full(filename, rebuild)
}
\arguments{
\item{filename}{The normalized DSS file path.}

\item{rebuild}{If TRUE, rebuild the catalog. For both condensed and
full catalog stores, the existing catalog will be dropped.
Additionally, for full catalog stores the \code{rebuild} parameter will
be passed to the \code{getCatalogedPathnames()} Java method.}
}
\value{
A character vector of pathnames.
}
\description{
DSS Catalog Retrieval
}
\keyword{internal}
