% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.r
\name{dss_close}
\alias{dss_close}
\alias{dss_close_all}
\title{Close DSS File}
\usage{
dss_close(filename)

dss_close_all()
}
\arguments{
\item{filename}{A DSS file path.}
}
\description{
Close a DSS file. This should only be used when the user has
finished working with a DSS file, and is not strictly required
as the package will implictly close a DSS file after every
operation.
}
