% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.r
\name{dss_require}
\alias{dss_require}
\alias{dss_connect}
\alias{dss_connect_win}
\alias{dss_connect_unix}
\title{Connect To DSS}
\usage{
dss_require(
  dss_home = getOption("dss.home"),
  message_level = getOption("dss.messagelevel")
)

dss_connect(
  dss_home = getOption("dss.home"),
  message_level = getOption("dss.messagelevel")
)

dss_connect_win(dss_home, message_level)

dss_connect_unix(dss_home, message_level)
}
\arguments{
\item{dss_home}{The path to the HEC-DSS install folder.}

\item{message_level}{An integer specifying the DSS message level.
Default is 2L. Use 1L to suppress all messages from DSS.}
}
\description{
Connect to DSS via the JVM.
}
\details{
both \code{\link[=dss_connect]{dss_connect()}} and \code{\link[=dss_require]{dss_require()}} configure the JVM
to use the HEC-DSS Java library. \code{\link[=dss_require]{dss_require()}} is designed for
use inside other functions; it returns \verb{FALSE`` and gives a warning (rather than an error as }\code{\link[=dss_connect]{dss_connect()}} does by default)
if the HEC-DSS Java library cannot be loaded correctly.
}
\keyword{internal}
