% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.r
\name{dss_connect_unix}
\alias{dss_connect_unix}
\title{Connect To DSS (Unix)}
\usage{
dss_connect_unix(dss_home, message_level, monolith)
}
\arguments{
\item{dss_home}{The path to the HEC-DSS install folder.}

\item{message_level}{An integer specifying the DSS message level.
Default is 2L. Use 1L to suppress all messages from DSS.}

\item{monolith}{If \code{TRUE}, connect to HEC-Monolith libraries.}
}
\description{
Connect to DSS in a Unix environment.
}
\keyword{internal}
