% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.r
\name{dss_connect_win}
\alias{dss_connect_win}
\title{Connect To DSS (Windows)}
\usage{
dss_connect_win(dss_home, message_level, monolith)
}
\arguments{
\item{dss_home}{The path to the HEC-DSS install folder.}

\item{message_level}{An integer specifying the DSS message level.
Default is 2L. Use 1L to suppress all messages from DSS.}
}
\description{
Connect to DSS in a Windows environment.
}
\keyword{internal}
