% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.r
\name{dss_convert}
\alias{dss_convert}
\title{DSS File Conversion}
\usage{
dss_convert(file, to)
}
\arguments{
\item{file}{A DSS file handle, e.g., output of \code{\link[=dss_open]{dss_open()}}.}

\item{to}{The destination file path.}
}
\description{
Convert between DSS file versions.
}
\seealso{
\code{\link[=dss_version]{dss_version()}} \code{\link[=dss_create]{dss_create()}}
}
