% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.r
\name{dss_create}
\alias{dss_create}
\title{Create Empty DSS File.}
\usage{
dss_create(filename, version = 7L)
}
\arguments{
\item{filename}{A DSS file path.}

\item{version}{The DSS version of the created file (6 or 7).}
}
\description{
Create an empty DSS file.
}
\seealso{
\code{\link[=dss_convert]{dss_convert()}}
}
