% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.r
\name{dss_delete}
\alias{dss_delete}
\title{Delete DSS Records}
\usage{
dss_delete(file, path, squeeze = FALSE)
}
\arguments{
\item{file}{A DSS file handle, e.g., output of \code{\link[=dss_open]{dss_open()}}.}

\item{path}{A vector of DSS paths.}

\item{squeeze}{If \code{TRUE}, squeeze the file after deleting records.}
}
\description{
Delete records from a DSS file.
}
\seealso{
\code{\link[=dss_open]{dss_open()}} \code{\link[=dss_catalog]{dss_catalog()}} \code{\link[=dss_squeeze]{dss_squeeze()}}
}
