% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.r
\name{dss_file}
\alias{dss_file}
\title{Open DSS File}
\usage{
dss_file(filename, ...)
}
\arguments{
\item{filename}{A valid filepath.}
}
\value{
A DSS file handle Java object.
}
\description{
Call the Java HecDss Open method.
}
\keyword{internal}
