% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.r
\name{dss_message_level}
\alias{dss_message_level}
\title{DSS Message Level}
\usage{
dss_message_level(message_level)
}
\arguments{
\item{message_level}{An integer specifying the DSS message level.
Default is 2L. Use 1L to suppress all messages from DSS.}
}
\description{
Set the DSS message level.
}
