% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.r
\name{dss_open}
\alias{dss_open}
\alias{dss_create}
\title{DSS File Handle}
\usage{
dss_open(filename)

dss_create(filename)
}
\arguments{
\item{filename}{Location of DSS file to open.}
}
\value{
A \code{hec.heclib.dss.HecDss} Java object reference.
}
\description{
Open or create a DSS file.
}
\details{
Objects created by this function must be explicitly closed
using the \verb{$close()} or \verb{$done()} methods. If this function fails
with a message that \code{javaHeclib.dll} cannot be found, this usually
indicates that \code{dss.home} is not set correctly.
}
