% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.r
\name{dss_parts_combine}
\alias{dss_parts_combine}
\title{DSS Parts To Path}
\usage{
dss_parts_combine(parts)
}
\arguments{
\item{parts}{A dataframe of DSS path parts "A", "B", "C",
"D", "E", "F".}
}
\value{
A vector of paths.
}
\description{
Combine DSS parts into a path.
}
