% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.r
\name{dss_parts_replace}
\alias{dss_parts_replace}
\title{Replace DSS Path Parts}
\usage{
dss_parts_replace(path, replacement = list())
}
\arguments{
\item{path}{A vector of DSS paths.}

\item{replacement}{A list of replacement path parts "A", "B", "C",
"D", "E", "F".}
}
\value{
A vector of paths, with parts replaced.
}
\description{
Replace DSS path parts with specified values.
}
