% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.r
\name{dss_parts_split}
\alias{dss_parts_split}
\title{DSS Path To Parts}
\usage{
dss_parts_split(path)
}
\arguments{
\item{path}{A vector of DSS paths.}
}
\value{
A dataframe with fields "path", "A", "B", "C",
"D", "E", "F".
}
\description{
Split DSS path into parts.
}
