% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.r
\name{dss_paths}
\alias{dss_paths}
\title{Get DSS Paths}
\usage{
dss_paths(file, pattern = "*", condensed = TRUE, rebuild = FALSE)
}
\arguments{
\item{file}{A DSS file handle, e.g., output of \code{\link[=dss_open]{dss_open()}}.}

\item{pattern}{character string containing a \link[base]{regular expression}
    (or character string for \code{fixed = TRUE}) to be matched
    in the given character vector.  Coerced by
    \code{\link[base]{as.character}} to a character string if possible.  If a
    character vector of length 2 or more is supplied, the first element
    is used with a warning.  Missing values are allowed except for
    \code{regexpr}, \code{gregexpr} and \code{regexec}.}

\item{condensed}{If \code{TRUE} returned the condensed path catalog.}

\item{rebuild}{If \code{TRUE}, force rebuild the catalog.}
}
\value{
A vector of paths.
}
\description{
Get a list of DSS paths in a file.
}
