% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.r
\name{dss_read_grid}
\alias{dss_read_grid}
\title{DSS Grid Data}
\usage{
dss_read_grid(gridObj)
}
\arguments{
\item{gridObj}{A \code{hec.io.GridContainer} Java object reference.}
}
\value{
A data frame.
}
\description{
Read DSS grid.
}
\keyword{internal}
