% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.r
\name{dss_read_paired}
\alias{dss_read_paired}
\title{DSS Paired Data}
\usage{
dss_read_paired(pdObj)
}
\arguments{
\item{pdObj}{A \code{hec.io.PairedDataContainer} Java object reference.}
}
\value{
A data frame.
}
\description{
Read paired data from a DSS file.
}
\keyword{internal}
