% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.r
\name{dss_read_timeseries}
\alias{dss_read_timeseries}
\title{DSS Time Series}
\usage{
dss_read_timeseries(tsObj, offset = FALSE)
}
\arguments{
\item{tsObj}{A \code{hec.io.TimeSeriesContainer} Java object reference.}

\item{offset}{(TimeSeries only) If \code{TRUE}, and values are period
averages, offset timestamps to center on periods.}
}
\value{
A data frame.
}
\description{
Read DSS time series.
}
\keyword{internal}
