% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.r
\name{dss_squeeze}
\alias{dss_squeeze}
\title{Squeeze a DSS File}
\usage{
dss_squeeze(file)
}
\arguments{
\item{file}{A DSS file handle, e.g., output of \code{\link[=dss_open]{dss_open()}}.}
}
\description{
Calls the squeeze method on file to remove deleted or overwritten
data.
}
