% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{dss_times}
\alias{dss_times}
\title{DSS Time Vector}
\usage{
dss_times(times, metadata, offset)
}
\arguments{
\item{times}{A vector of DSS times.}

\item{metadata}{The metadata associated with \code{obj}, i.e.,
output of \code{\link[=java_metadata]{java_metadata()}}.}
}
\value{
A POSIXct vector.
}
\description{
Extract timestamps from a DSS object.
}
\keyword{internal}
