% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.r
\name{dss_times_from_posix}
\alias{dss_times_from_posix}
\title{DSS Times From POSIXct}
\usage{
dss_times_from_posix(datetimes, dss_interval)
}
\arguments{
\item{datetimes}{A POSIXct vector.
Default is 60L (one  minute).}

\item{dss_interval}{The DSS time interval, e.g., "15MIN", "1DAY",
"IR-CENTURY", etc.}
}
\value{
A named list for writing to DSS.
}
\description{
Convert POSIXct timestamps to DSS times.
}
\keyword{internal}
