% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{dss_timezone}
\alias{dss_timezone}
\title{DSS Time Zone}
\usage{
dss_timezone(tzoffset)
}
\arguments{
\item{tzoffset}{The DSS object's "timeZoneRawOffset" attribute
(time offset in milliseconds).}
}
\value{
An R timezone string in format "Etc/GMT*".
}
\description{
Convert DSS raw timezone offset to an R timezone.
}
\keyword{internal}
