% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.r
\name{dss_to_grid}
\alias{dss_to_grid}
\title{Grid to DSS}
\usage{
dss_to_grid(r)
}
\arguments{
\item{r}{A raster.}
}
\value{
A \code{hec.io.GridContainer} Java object reference.
}
\description{
Convert a raster to a DSS Grid object.
}
\keyword{internal}
