% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.r
\name{dss_to_paired}
\alias{dss_to_paired}
\title{Paired Data to DSS}
\usage{
dss_to_paired(d, param_part)
}
\arguments{
\item{d}{A data frame.}
}
\value{
A \code{PairedDataContainer} Java object reference.
}
\description{
Convert a data frame to a DSS Paired Data object.
}
\keyword{internal}
