% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.r
\name{dss_version}
\alias{dss_version}
\title{DSS File Version}
\usage{
dss_version(file)
}
\arguments{
\item{file}{A DSS file handle, e.g., output of \code{\link[=dss_open]{dss_open()}}.}
}
\value{
An integer
}
\description{
Get the DSS file version.
}
\seealso{
\code{\link[=dss_convert]{dss_convert()}}
}
