% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.r
\name{dss_write}
\alias{dss_write}
\title{Write to DSS}
\usage{
dss_write(x, file, path, attributes = list())
}
\arguments{
\item{x}{The dataframe (Time Series or Paired Data) or raster (Grid)
to write to DSS.}

\item{file}{A DSS file handle, e.g., output of \code{\link[=dss_open]{dss_open()}}.}

\item{attributes}{A named list of DSS attributes.}
}
\description{
Write to DSS
}
