% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\docType{package}
\name{dssrip}
\alias{dssrip}
\title{dssrip}
\description{
A wrapper for the
\href{https://www.hec.usace.army.mil/confluence/dssjavaprogrammer}{HEC-DSS Java API}
allowing an interface between R and many common hydrologic models.
This is a rewrite of the original "dssrip" package.
}
