% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\docType{package}
\name{dssrip2-package}
\alias{dssrip2}
\alias{dssrip2-package}
\title{dssrip2}
\description{
A wrapper for the
\href{https://www.hec.usace.army.mil/confluence/dssjavaprogrammer}{HEC-DSS Java API}
allowing an interface between R and many common hydrologic models.
This is a rewrite of the original "dssrip" package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mkoohafkan/dssrip2}
  \item \url{https://hydroecology.net/dssrip2/}
  \item Report bugs at \url{https://github.com/mkoohafkan/dssrip2/issues}
}

}
\author{
\strong{Maintainer}: Michael Koohafkan \email{michael.c.koohafkan@usace.army.mil}

Authors:
\itemize{
  \item Evan Heisman \email{evan.a.heisman@usace.army.mil}
  \item Cameron Bracken
}

}
