% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{format_datetimes}
\alias{format_datetimes}
\title{Format Datetimes}
\usage{
format_datetimes(x)
}
\arguments{
\item{x}{A vector to be coerced to datetimes}
}
\value{
A vector of datetimes.
}
\description{
Helper function to format datetimes.
}
\keyword{internal}
