% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{format_na}
\alias{format_na}
\title{Format NA Values}
\usage{
format_na(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The vector \code{x}, possibly with NA values
}
\description{
Replace the DSS placeholder NA value with R NA values.
}
\keyword{internal}
