% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java-utils.r
\name{get_field}
\alias{get_field}
\alias{use_method}
\title{Java Object Fields And Methods}
\usage{
get_field(jObject, field)

use_method(jObject, method, ...)
}
\arguments{
\item{jObject}{A Java object.}

\item{shortname}{The Java field name.}
}
\value{
A list of metadata.
}
\description{
Helper functions to get field or call methods of a Java object.
}
\keyword{internal}
