% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.r
\name{get_jars}
\alias{get_jars}
\title{Get DSS Jars}
\usage{
get_jars(dss_home, monolith)
}
\arguments{
\item{dss_home}{The path to the DSS install.}
}
\value{
A list of paths to required jars.
}
\description{
Get paths to required DSS jars.
}
\keyword{internal}
