% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.r
\name{guess_dss_container}
\alias{guess_dss_container}
\title{Guess DSS Class}
\usage{
guess_dss_container(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
The DSS Java object type.
}
\description{
Guess the DSS container based on the R object structure.
}
\keyword{internal}
