% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.r
\name{java_fields}
\alias{java_fields}
\title{Fields Data Frame}
\usage{
java_fields(jObject)
}
\arguments{
\item{jObject}{A Java object.}
}
\value{
A dataframe with fields "FULLNAME", "SHORTNAME", "CLASS",
and "SIGNATURE".
}
\description{
Get a drataframe of Java object fields.
}
\keyword{internal}
