% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.r
\name{java_metadata}
\alias{java_metadata}
\title{Get Java Metadata}
\usage{
java_metadata(jObject)
}
\arguments{
\item{jObject}{A Java object.}
}
\value{
A named list of metadata.
}
\description{
Get a list of fields in the java object and their values.
}
