% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{java_to_na}
\alias{java_to_na}
\alias{na_to_java}
\title{Convert NA Values}
\usage{
java_to_na(x)

na_to_java(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The vector \code{x}, possibly with NA/placeholder values.
}
\description{
Convert between DSS placeholder and R NA values.
}
\keyword{internal}
