% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monolith.r
\name{maven-helpers}
\alias{maven-helpers}
\alias{maven_download_url}
\alias{maven_path}
\alias{maven_sha1}
\title{Maven Helpers}
\usage{
maven_download_url(id, extension, url = "https://repo.maven.apache.org/maven2")

maven_path(id, extension)

maven_sha1(id, extension, url = "https://repo.maven.apache.org/maven2")
}
\arguments{
\item{id}{The asset ID.}

\item{extension}{the Asset file extension.}

\item{The}{base repository URL.}
}
\description{
Helper functions for searching the Maven central repository.
}
\keyword{internal}
