% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.r
\name{metafield}
\alias{metafield}
\title{Get Metadata Field}
\usage{
metafield(jObject, shortname, signature)
}
\arguments{
\item{jObject}{A Java object.}

\item{shortname}{The Java object shortname.}

\item{signature}{The Java object signature.}
}
\value{
A list of metadata.
}
\description{
Helper function to get metadata field value.
}
\keyword{internal}
