% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monolith.r
\name{monolith_assets}
\alias{monolith_assets}
\title{Monolith Assets}
\usage{
monolith_assets(requirements_file, stop_on_fail = TRUE)
}
\arguments{
\item{requirements_file}{A YAML file listing required assets.}

\item{stop_on_fail}{If \code{TRUE}, fail if any assets cannot be found.}
}
\value{
A data frame of required assets.
}
\description{
Search for assets listed in a YAML file.
}
\keyword{internal}
