% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monolith.r
\name{monolith_default_dir}
\alias{monolith_default_dir}
\title{HEC Monolith Default Directory}
\usage{
monolith_default_dir()
}
\description{
Platform-dependent default install directory for HEC-Monolith.
}
\keyword{internal}
