% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monolith.r
\name{monolith_requirements}
\alias{monolith_requirements}
\title{Monolith Requirements}
\usage{
monolith_requirements(requirements_file)
}
\arguments{
\item{requirements_file}{A YAML file listing required assets.}
}
\value{
A data frame of required assets.
}
\description{
Read a list of required assets from a YAML file.
}
\keyword{internal}
