% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{na_fill}
\alias{na_fill}
\title{Fill NA Values}
\usage{
na_fill(v, forward = TRUE)
}
\arguments{
\item{v}{A vector.}

\item{forward}{If \code{TRUE}, fill starting from first to last element,
If \code{FALSE}, fill in from last to first element.}
}
\value{
The vector \code{v}, with NA values filled.
}
\description{
Fill NA values with the previous (or following) value.
}
