% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.r
\name{normalize_path}
\alias{normalize_path}
\title{Normalize DSS File Path}
\usage{
normalize_path(filename, exists)
}
\arguments{
\item{filename}{A DSS file path.}

\item{exists}{If \code{TRUE}, assert that the DSS file exists.}
}
\value{
A normalized file path.
}
\description{
Normalize a DSS file path. A wrapper for \code{\link[base:normalizePath]{base::normalizePath()}}
with winslash set to \code{"/"}.
}
\keyword{internal}
