% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.r
\name{as_hectime}
\alias{as_hectime}
\title{POSIXct to HecTime}
\usage{
as_hectime(x, granularity_seconds)
}
\arguments{
\item{x}{A POSIXct timestamp.}

\item{granularity_seconds}{The minimum time interval, in seconds.}
}
\value{
A Java \code{hec.heclib.util.HecTime} object.
}
\description{
Convert an R POSIXct datetime to Java HecTime.
}
\keyword{internal}
