% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.r
\name{dss-assertions}
\alias{dss-assertions}
\alias{assert_dss_file}
\alias{assert_path_format}
\alias{assert_parts_format}
\alias{assert_read_support}
\alias{assert_write_support}
\alias{assert_timeseries}
\alias{assert_paired}
\alias{assert_grid}
\title{DSS Assertions}
\usage{
assert_dss_file(file)

assert_path_format(path)

assert_parts_format(parts)

assert_read_support(obj)

assert_write_support(obj)

assert_timeseries(obj)

assert_paired(obj)

assert_grid(obj)
}
\description{
Helper functions for DSS type and format checking.
}
\section{Functions}{
\itemize{
\item \code{assert_dss_file()}: Assert the supplied object is a DSS file handle.

\item \code{assert_path_format()}: Assert the supplied path is in the correct format.

\item \code{assert_parts_format()}: Assert set of supplied parts is named and complete.

\item \code{assert_read_support()}: Assert the supplied rJava object is a DSS timeseries.

\item \code{assert_write_support()}: Assert the supplied rJava object is a DSS timeseries.

\item \code{assert_timeseries()}: Assert the supplied rJava object is a DSS timeseries.

\item \code{assert_paired()}: Assert the supplied rJava object is a DSS paired data container.

\item \code{assert_grid()}: Assert the supplied rJava object is a DSS grid data container.

}}
\keyword{internal}
