% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.r
\name{dss_attributes}
\alias{dss_attributes}
\title{Read DSS Attributes}
\usage{
dss_attributes(file, path)
}
\arguments{
\item{file}{A DSS file handle, e.g., output of \code{\link[=dss_open]{dss_open()}}.}

\item{path}{The DSS path to query attributes for.}
}
\value{
A List of attributes.
}
\description{
Read attributes from a DSS file.
}
\seealso{
\code{\link[=dss_read]{dss_read()}}
}
