% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.r
\name{dss_catalog}
\alias{dss_catalog}
\title{Get DSS Paths}
\usage{
dss_catalog(file, pattern = "*", condensed = TRUE, rebuild = FALSE)
}
\arguments{
\item{file}{A DSS file handle, e.g., output of \code{\link[=dss_open]{dss_open()}}.}

\item{pattern}{character string containing a regular expression to
be matched against the DSS pathname catalog.}

\item{condensed}{If \code{TRUE} returned the condensed path catalog.}

\item{rebuild}{If \code{TRUE}, force rebuild the catalog.}
}
\value{
A vector of paths.
}
\description{
Get a list of DSS paths in a file.
}
\seealso{
\code{\link[=dss_open]{dss_open()}} \code{\link[=dss_read]{dss_read()}} \code{\link[=dss_parts_split]{dss_parts_split()}}
\code{\link[=dss_parts_combine]{dss_parts_combine()}} \code{\link[=dss_parts_replace]{dss_parts_replace()}}
}
