% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monolith.r
\name{dss_install_monolith}
\alias{dss_install_monolith}
\alias{dss_install_monolith_win}
\alias{dss_install_monolith_unix}
\title{Install Monolith}
\usage{
dss_install_monolith(install_path, overwrite = TRUE, requirements_file)

dss_install_monolith_win(assets, install_path, overwrite)

dss_install_monolith_unix(assets, install_path, overwrite)
}
\arguments{
\item{install_path}{The directory to download the monolith libraries to.}

\item{overwrite}{If \code{TRUE}, delete any existing data in
\code{install_path} and recreate.}

\item{requirements_file}{The list of required assets. For expert use only.}
}
\description{
Download the HEC Monolith Libraries.
}
\seealso{
\code{\link[=dss_connect]{dss_connect()}}
}
\keyword{internal}
