% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.r
\name{dss_open}
\alias{dss_open}
\alias{dss_create}
\title{DSS File Handle}
\usage{
dss_open(filename)

dss_create(filename, version = 7L)
}
\arguments{
\item{filename}{Location of DSS file to open.}

\item{version}{The DSS version of the created file (6 or 7).}
}
\value{
A \code{hec.heclib.dss.HecDss} Java object reference.
}
\description{
Open or create a DSS file.
}
\details{
Objects created by this function must be explicitly closed
using the \verb{$close()} or \verb{$done()} methods. If this function fails
with a message that \code{javaHeclib.dll} cannot be found, this usually
indicates that \code{dss.home} is not set correctly.
}
\seealso{
\code{\link[=dss_close]{dss_close()}} \code{\link[=dss_version]{dss_version()}}
}
