% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.r
\name{dss_parts_split}
\alias{dss_parts_split}
\title{DSS Path To Parts}
\usage{
dss_parts_split(path, keep = FALSE)
}
\arguments{
\item{path}{A vector of DSS paths.}

\item{keep}{If \code{TRUE}, preserve the supplied paths in the output.}
}
\value{
A dataframe with fields "A", "B", "C",
"D", "E", "F" and optional field "path" (if \code{keep = TRUE}).
}
\description{
Split DSS path into parts.
}
\seealso{
\code{\link[=dss_catalog]{dss_catalog()}} \code{\link[=dss_parts_combine]{dss_parts_combine()}} \code{\link[=dss_parts_replace]{dss_parts_replace()}}
}
