% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.r
\name{dss_read}
\alias{dss_read}
\title{Read DSS Data}
\usage{
dss_read(file, path, full = TRUE, offset = FALSE)
}
\arguments{
\item{file}{A DSS file handle, e.g., output of \code{\link[=dss_open]{dss_open()}}.}

\item{path}{The DSS path to extract.}

\item{full}{If \code{TRUE}, read the entire dataset. If \code{FALSE}, only
read the section as specified in the D Part of the path.}

\item{offset}{(TimeSeries only) If \code{TRUE}, and values are period
averages, offset timestamps to center on periods.}
}
\value{
A data frame. For a time series object, the first column
will be "datetime". Further columns will be named according to
the DSS "parameter" value specified in the object metadata.
}
\description{
Read data from a DSS file.
}
\details{
The returned data frame includes additional attributes
\code{"dss.type"} and \code{"dss.units"} describing the DSS data format.
}
\seealso{
\code{\link[=dss_attributes]{dss_attributes()}}
}
