% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.r
\name{dss_read}
\alias{dss_read}
\title{Read DSS Records}
\usage{
dss_read(file, path, full = TRUE)
}
\arguments{
\item{file}{A DSS file handle, e.g., output of \code{\link[=dss_open]{dss_open()}}.}

\item{path}{The DSS path to extract.}

\item{full}{If \code{TRUE}, read the entire dataset. If \code{FALSE}, only
read the section as specified in the D Part of the path.}
}
\value{
A data frame. For a time series object, the first column
will be "datetime". Further columns will be named according to
the DSS "parameter" value specified in the object metadata.
}
\description{
Read records from a DSS file.
}
\details{
The returned data frame includes additional attribute
\code{"dss_attributes"} describing the DSS data format.
}
\seealso{
\code{\link[=dss_open]{dss_open()}} \code{\link[=dss_catalog]{dss_catalog()}} \code{\link[=dss_attributes]{dss_attributes()}}
}
