% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.r
\name{dss_squeeze}
\alias{dss_squeeze}
\title{Squeeze a DSS File}
\usage{
dss_squeeze(filename)
}
\arguments{
\item{filename}{A DSS file path.}
}
\description{
Calls the squeeze method on file to remove deleted or overwritten
data.
}
\seealso{
\code{\link[=dss_delete]{dss_delete()}}
}
