% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{dss_times}
\alias{dss_times}
\title{DSS Time Vector}
\usage{
dss_times(times, info, offset)
}
\arguments{
\item{times}{A vector of DSS times.}

\item{info}{A list of attributes relevant to processing timestamps.
Must include "timeGranularitySeconds", "timeIntervalSeconds",
"timeZoneRawOffset", and "type".}
}
\value{
A POSIXct vector.
}
\description{
Extract timestamps from a DSS object.
}
\keyword{internal}
