% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.r
\name{dss_times_from_posix}
\alias{dss_times_from_posix}
\title{DSS Times From POSIXct}
\usage{
dss_times_from_posix(datetimes)
}
\arguments{
\item{datetimes}{A POSIXct vector.}

\item{type}{The DSS data type, e.g. "PER_AVER" or "INST".}

\item{granularity}{The smallest time interval in the data.
Default is 60L (one  minute).}
}
\value{
A named list for writing to DSS.
}
\description{
Convert POSIXct timestamps to DSS times.
}
\keyword{internal}
