% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.r
\name{dss_to_timeseries}
\alias{dss_to_timeseries}
\title{DSS to Timeseries}
\usage{
dss_to_timeseries(d, attributes)
}
\arguments{
\item{datetimes}{A POSIXct vector.}

\item{type}{The DSS data type, e.g. "PER_AVER" or "INST".}

\item{granularity}{The smallest time interval in the data.
Default is 60L (one  minute).}
}
\value{
A named list for writing to DSS.
}
\description{
Convert a data frame to a DSS Time Series object.
}
\keyword{internal}
