% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{format_datetimes}
\alias{format_datetimes}
\title{Format Datetimes}
\usage{
format_datetimes(x, default_tz = "etc/GMT+0")
}
\arguments{
\item{x}{A vector to be coerced to datetimes}

\item{default_tx}{The default timezone to use for coercing
objects without a specified timezone.}
}
\value{
A vector of datetimes.
}
\description{
Helper function to format datetimes.
}
\keyword{internal}
