% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.r
\name{list_attributes}
\alias{list_attributes}
\alias{list_attributes_timeseries}
\alias{list_attributes_paired}
\alias{list_attributes_grid}
\title{List DSS Attributes}
\usage{
list_attributes(class_name)

list_attributes_timeseries()

list_attributes_paired()

list_attributes_grid()
}
\arguments{
\item{class_name}{A DSS Java object class name.}
}
\value{
A list of required and optional attribute names.
}
\description{
Return a vector of attribute names associated with a
DSS object class.
}
\keyword{internal}
